pyproject.toml
[tool.poetry]
name = "scraper-service"
version = "0.1.0"
description = "UniBo Scraper Service"
authors = ["UniBo Toolkit"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.35"}
asyncpg = "^0.30.0"
redis = {extras = ["hiredis"], version = "^5.2.0"}
python-dotenv = "^1.0.1"
coloredlogs = "^15.0.1"
pydantic = "^2.9.2"
pydantic-settings = "^2.6.1"
apscheduler = "^3.10.4"
unibo-toolkit = "0.4.0"
httpx = "^0.28.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
black = "^24.10.0"
isort = "^5.13.2"
ruff = "^0.7.4"
mypy = "^1.13.0"
sqlacodegen = "^3.0.0rc5"
psycopg2-binary = "^2.9.11"

[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "PT"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
